COS-27: правки
