+++ packages/client/src/shared/constants/index.tsx
@@ -126,7 +126,6 @@ export enum RoutesName {
   LOGIN = "/login",
   REGISTRATION = "/registration",
   FORUM_DETAIL = "/forum/:id",
-  OAUTH = "/oauth",
 }
 
 export const MenuItems: MenuLinkType[] = [
